+++ package.json
@@ -24,20 +24,20 @@
   },
   "scripts": {
     "clean": "rm -rf dist",
-    "build": "npm run clean; tsc",
-    "test": "TZ=utc mocha"
+    "build": "npm run clean; tsc --project tsconfig.dist.json",
+    "test": "env TS_NODE_PROJECT=\"tsconfig.test.json\" TZ=utc jest"
   },
   "devDependencies": {
     "@types/assert": "^1.5.6",
-    "@types/mocha": "^9.1.1",
+    "@types/jest": "^29.2.2",
     "assert": "^2.0.0",
-    "mocha": "^10.0.0",
-    "ts-node": "^10.9.1",
+    "jest": "^29.3.1",
+    "ts-jest": "^29.0.3",
     "typescript": "^4.5.5"
   },
   "files": [
     "dist/**/*"
   ],
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts"
+  "main": "dist/matter.js",
+  "types": "dist/matter.d.ts"
 }
