--- src/index.ts
@@ -1,5 +1,0 @@
-/**
- * @license
- * Copyright 2022 Project CHIP Authors
- * SPDX-License-Identifier: Apache-2.0
- */
