+++ src/tlv/TlvFloat.ts
@@ -4,11 +4,11 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { DataReaderLE } from "../util/DataReaderLE";
-import { DataWriterLE } from "../util/DataWriterLE";
-import { FLOAT32_MAX, FLOAT32_MIN } from "../util/Number";
-import { TlvType, TlvCodec, TlvTag, TlvLength, TlvTypeLength } from "./TlvCodec";
-import { TlvSchema } from "./TlvSchema";
+import { DataReaderLE } from "../util/DataReaderLE.js";
+import { DataWriterLE } from "../util/DataWriterLE.js";
+import { FLOAT32_MAX, FLOAT32_MIN } from "../util/Number.js";
+import { TlvType, TlvCodec, TlvTag, TlvLength, TlvTypeLength } from "./TlvCodec.js";
+import { TlvSchema } from "./TlvSchema.js";
 
 /**
  * Schema to encode a float in TLV.
