+++ src/tlv/TlvSchema.ts
@@ -4,11 +4,11 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { ByteArray } from "../util/ByteArray";
-import { DataReaderLE } from "../util/DataReaderLE";
-import { DataWriterLE } from "../util/DataWriterLE";
-import { Schema } from "../util/schema/Schema";
-import { TlvTag } from "./TlvCodec";
+import { ByteArray } from "../util/ByteArray.js";
+import { DataReaderLE } from "../util/DataReaderLE.js";
+import { DataWriterLE } from "../util/DataWriterLE.js";
+import { Schema } from "../util/schema/Schema.js";
+import { TlvTag } from "./TlvCodec.js";
 
 export abstract class TlvSchema<T> extends Schema<T, ByteArray> {
     /** @override */
