+++ test-web/index.html
@@ -1,0 +1,9 @@
+<html>
+    <header>
+        <script type="module">
+            import { TlvString } from "../dist/matter.js";
+
+            console.log(TlvString().encode("my string").toHex());
+        </script>
+    </header>
+</html>
