+++ tsconfig.base.json
@@ -1,0 +1,9 @@
+/* Base configuration */
+{
+  "compilerOptions": {
+    "target": "ES2015",
+    "module": "ES2015",
+    "noImplicitAny": true,
+    "strict": true
+  }
+}
