+++ tsconfig.test.json
@@ -1,9 +1,9 @@
-/* Dev configuration */
+/* Configuration for mocha unit tests */
 {
   "extends": "./tsconfig.base.json",
   "compilerOptions": {
+    "esModuleInterop": true, // Requires to import assert
     "allowSyntheticDefaultImports": true, // Used to import assert in unit tests
   },
-  /* During dev, src are included so they are also modified during refactorings. */
   "include": ["src/**/*.ts", "test-node/**/*.ts"],
 }
